requirements.txt
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
bleach==6.2.0
bokeh==3.7.3
boto3==1.38.22
botocore==1.38.22
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
fonttools==4.58.4
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-billing==1.16.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
httplib2==0.22.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
linkify-it-py==2.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
panel==1.7.1
param==2.2.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
tabulate==0.9.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
webencodings==0.5.1
xyzservices==2025.4.0
